Gathering serialization and deserialization of CityRoads in routing.
