+++ routing/city_roads_loader.hpp
@@ -3,7 +3,6 @@
 #include "indexer/mwm_set.hpp"
 
 #include "coding/memory_region.hpp"
-#include "coding/writer.hpp"
 
 #include <cstdint>
 #include <memory>
@@ -15,32 +14,6 @@ class DataSource;
 
 namespace routing
 {
-struct CityRoadsHeader
-{
-  template <class Sink>
-  void Serialize(Sink & sink) const
-  {
-    WriteToSink(sink, m_version);
-    WriteToSink(sink, m_endianness);
-    WriteToSink(sink, m_dataSize);
-  }
-
-  template <class Source>
-  void Deserialize(Source & src)
-  {
-    m_version = ReadPrimitiveFromSource<uint16_t>(src);
-    m_endianness = ReadPrimitiveFromSource<uint16_t>(src);
-    m_dataSize = ReadPrimitiveFromSource<uint32_t>(src);
-  }
-
-  uint16_t m_version = 0;
-  // Field |m_endianness| is reserved for endianness of the section.
-  uint16_t m_endianness = 0;
-  uint32_t m_dataSize = 0;
-};
-
-static_assert(sizeof(CityRoadsHeader) == 8, "Wrong header size of city_roads section.");
-
 class CityRoadsLoader
 {
 public:
